requirements.txt
asgiref==3.7.2
awsebcli==3.20.9
beautifulsoup4==4.11.1
botocore==1.31.45
cement==2.8.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
colorama==0.4.3
cryptography==39.0.0
defusedxml==0.7.1
deprecation==2.1.0
distlib==0.3.4
Django==4.1.4
django-autoslug==1.9.8
django-bootstrap4==22.2
django-ckeditor==6.5.1
django-cors-headers==4.2.0
django-countries==7.5.1
django-crispy-forms==1.14.0
django-email-verification==0.3.1
django-etc==1.4.0
django-filter==22.1
django-hitcount==1.3.5
django-js-asset==2.0.0
django-paypal==2.0
django-tagging-autocomplete==0.5.1
django-widget-tweaks==1.4.12
djangorestframework==3.14.0
filelock==3.7.1
idna==3.4
jmespath==1.0.1
mysql-connector==2.2.9
mysql-connector-python==8.0.29
oauthlib==3.2.2
packaging==23.0
pathspec==0.10.1
Pillow==9.3.0
platformdirs==2.5.2
protobuf==4.21.2
py3dns==3.2.1
pycparser==2.21
PyJWT==2.6.0
pypiwin32==223
python-dateutil==2.8.2
python-decouple==3.6
python3-openid==3.2.0
pytz==2022.7.1
pyuca==1.2
pywin32==306
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
semantic-version==2.8.5
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
stripe==5.0.0
termcolor==1.1.0
typing_extensions==4.5.0
tzdata==2022.1
urllib3==1.26.14
validate-email==1.3
virtualenv==20.15.1
wcwidth==0.1.9
